+++ node_prog/node_prog_type.h
@@ -70,7 +70,7 @@ namespace node_prog
     class Packable 
     {
         public:
-            virtual size_t size() const  = 0;
+            virtual uint64_t size() const  = 0;
             virtual void pack(e::buffer::packer& packer) const = 0;
             virtual void unpack(e::unpacker& unpacker) = 0;
     };
